CDC #27 - Renaming PlaceResultsList to SearchResultsList; Updating to use more generic rendering function; Removing "react-instantsearch" dependency
