+++ packages/core-data/package.json
@@ -31,7 +31,6 @@
     "instantsearch.js": "^4.66.0",
     "lucide-react": "^0.321.0",
     "postcss-import": "^16.1.0",
-    "react-instantsearch": "^7.5.4",
     "react-virtualized-auto-sizer": "^1.0.22",
     "react-window": "^1.8.10",
     "typesense-instantsearch-adapter": "^2.8.0",
